armv8: Fix transfer list passing with KVM
When loading U-Boot in a low memory address, the adpr instruction
to preserve saved_args[] end up within the memory space of QEMUs
mapped pflashes. As a result the following store pair instructions
lead to a crash since instructions that modify more than one registers
can't trap into KVM.

This problem can be solved with loading U-Boot with -kernel
and defining a different text base address. But when running with
-bios we always start from 0x0. So let's postpone the calculation
of saved_args[] post relocation, where we will have an address that
doesn't collide with the QEMU flashes.

Acked-by: Raymond Mao <raymondmaoca@gmail.com>
Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>