arm: qemu: Eliminate fdt_high and initrd_high misuse
The fdt_high and initrd_high have nasty side-effects , which may lead
to DT placed at 4-byte aligned offset when used in place, which then
prevents Linux on arm64 from booting. This is difficult to debug and
inobvious, with little to no gain. Remove this to let U-Boot place the
DT at correctly aligned address.

Signed-off-by: Marek Vasut <marek.vasut+renesas@mailbox.org>