Gitlab: sage: Add Pine64+ platform
This adds the Pine64+ platform to the sage lab, for both legacy and lwIP
networking stacks. In order to build this platform we need to copy
certain files that were built in the container to /tmp and then set
BINMAN_INDIRS to /tmp in order to find them when building.

For now, we disable the test_net_pxe_boot_config test on lwIP as it
leads to a crash that needs to be investigated.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
Cc: Andre Przywara <andre.przywara@arm.com>
Cc: Jerome Forissier <jerome.forissier@linaro.org>

Andre, Jerome, how would you like to handle looking in to this issue?
https://source.denx.de/u-boot/u-boot/-/jobs/1303375#L2235 shows the
crash and I don't have enough sunxi targets to know if this is more
broadly breakable or not. But it doesn't happen on lwIP + Pi for
example, nor lwIP + Hummingboard (mx6cuboxi defconfig). I can't at this
point give remote access to the lab directly, but I can help come up
with a gitlab job that'll acquire the board, build and try and boot
whatever you have.